requirements.txt
# Core dependencies
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
langchain-experimental==0.0.49

# UI Framework
streamlit==1.30.0
streamlit-chat==0.1.1

# Data Analysis
pandas==2.1.4
numpy==1.24.4
scipy==1.11.4
scikit-learn==1.3.2
matplotlib==3.8.2
plotly==5.17.0
seaborn==0.13.0

# Web Tools
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.16.0
httpx==0.26.0

# File Processing
openpyxl==3.1.2
PyPDF2==3.0.1
python-docx==1.1.0

# API Integration
openai==1.55.3
tavily-python==0.3.3

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
typing-extensions>=4.11.0
tiktoken==0.5.2

# Development
pytest==7.4.4
black==23.12.1
isort==5.13.2
mypy==1.8.0

# Memory and Storage
chromadb==0.4.22
faiss-cpu==1.7.4

# Code execution safety
RestrictedPython==6.0

# Additional tools
wikipedia==1.4.0
duckduckgo-search==3.9.6
arxiv==2.1.0
